fix: address copilot review feedback for live-graph
- Break circular dependency between app.tsx and authoring.tsx by
  extracting baseAuthoringProps into authoring-config.ts
- Apply coerceToFiniteOrNull to fetched recording rows for consistency
  with live tick processing
- Add chartHeight validation in customValidate (must be positive finite)
